// test/integration/test/sql-builder/fixtures/generated/contract.d.ts
export type CodecTypes = PgTypes & PgVectorTypes;
export type LaneCodecTypes = CodecTypes;
export type QueryOperationTypes = PgAdapterQueryOps<CodecTypes> &
  PgVectorQueryOperationTypes<CodecTypes>;
// ...
export type FieldOutputTypes = {
  readonly public: {
    readonly Article: { readonly id: Char<36>; readonly title: CodecTypes['pg/text@1']['output'] };
    readonly Comment: {
      readonly id: CodecTypes['pg/int4@1']['output'];
      readonly body: CodecTypes['pg/text@1']['output'];
      readonly postId: CodecTypes['pg/int4@1']['output'];
    };
    readonly Post: {
      readonly id: CodecTypes['pg/int4@1']['output'];
      readonly title: CodecTypes['pg/text@1']['output'];
      readonly userId: CodecTypes['pg/int4@1']['output'];
      readonly views: CodecTypes['pg/int4@1']['output'];
      readonly embedding: Vector<3> | null;
    };
    readonly Profile: {
      readonly id: CodecTypes['pg/int4@1']['output'];
      readonly userId: CodecTypes['pg/int4@1']['output'];
      readonly bio: CodecTypes['pg/text@1']['output'];
    };
    readonly User: {
      readonly id: CodecTypes['pg/int4@1']['output'];
      readonly name: CodecTypes['pg/text@1']['output'];
      readonly email: CodecTypes['pg/text@1']['output'];
      readonly invitedById: CodecTypes['pg/int4@1']['output'] | null;
    };
  };
};
export type FieldInputTypes = {
  readonly public: {
    readonly Article: {
      readonly id: CodecTypes['sql/char@1']['input'];
      readonly title: CodecTypes['pg/text@1']['input'];
    };
    readonly Comment: {
      readonly id: CodecTypes['pg/int4@1']['input'];
      readonly body: CodecTypes['pg/text@1']['input'];
      readonly postId: CodecTypes['pg/int4@1']['input'];
    };
    readonly Post: {
      readonly id: CodecTypes['pg/int4@1']['input'];
      readonly title: CodecTypes['pg/text@1']['input'];
      readonly userId: CodecTypes['pg/int4@1']['input'];
      readonly views: CodecTypes['pg/int4@1']['input'];
      readonly embedding: CodecTypes['pg/vector@1']['input'] | null;
    };
    readonly Profile: {
      readonly id: CodecTypes['pg/int4@1']['input'];
      readonly userId: CodecTypes['pg/int4@1']['input'];
      readonly bio: CodecTypes['pg/text@1']['input'];
    };
    readonly User: {
      readonly id: CodecTypes['pg/int4@1']['input'];
      readonly name: CodecTypes['pg/text@1']['input'];
      readonly email: CodecTypes['pg/text@1']['input'];
      readonly invitedById: CodecTypes['pg/int4@1']['input'] | null;
    };
  };
};
export type StorageColumnTypes = {
  readonly public: {
    readonly articles: { readonly id: Char<36>; readonly title: CodecTypes['pg/text@1']['output'] };
    readonly comments: {
      readonly body: CodecTypes['pg/text@1']['output'];
      readonly id: CodecTypes['pg/int4@1']['output'];
      readonly post_id: CodecTypes['pg/int4@1']['output'];
    };
    readonly posts: {
      readonly embedding: Vector<3> | null;
      readonly id: CodecTypes['pg/int4@1']['output'];
      readonly title: CodecTypes['pg/text@1']['output'];
      readonly user_id: CodecTypes['pg/int4@1']['output'];
      readonly views: CodecTypes['pg/int4@1']['output'];
    };
    readonly profiles: {
      readonly bio: CodecTypes['pg/text@1']['output'];
      readonly id: CodecTypes['pg/int4@1']['output'];
      readonly user_id: CodecTypes['pg/int4@1']['output'];
    };
    readonly users: {
      readonly email: CodecTypes['pg/text@1']['output'];
      readonly id: CodecTypes['pg/int4@1']['output'];
      readonly invited_by_id: CodecTypes['pg/int4@1']['output'] | null;
      readonly name: CodecTypes['pg/text@1']['output'];
    };
  };
};
export type StorageColumnInputTypes = {
  readonly public: {
    readonly articles: {
      readonly id: CodecTypes['sql/char@1']['input'];
      readonly title: CodecTypes['pg/text@1']['input'];
    };
    readonly comments: {
      readonly body: CodecTypes['pg/text@1']['input'];
      readonly id: CodecTypes['pg/int4@1']['input'];
      readonly post_id: CodecTypes['pg/int4@1']['input'];
    };
    readonly posts: {
      readonly embedding: CodecTypes['pg/vector@1']['input'] | null;
      readonly id: CodecTypes['pg/int4@1']['input'];
      readonly title: CodecTypes['pg/text@1']['input'];
      readonly user_id: CodecTypes['pg/int4@1']['input'];
      readonly views: CodecTypes['pg/int4@1']['input'];
    };
    readonly profiles: {
      readonly bio: CodecTypes['pg/text@1']['input'];
      readonly id: CodecTypes['pg/int4@1']['input'];
      readonly user_id: CodecTypes['pg/int4@1']['input'];
    };
    readonly users: {
      readonly email: CodecTypes['pg/text@1']['input'];
      readonly id: CodecTypes['pg/int4@1']['input'];
      readonly invited_by_id: CodecTypes['pg/int4@1']['input'] | null;
      readonly name: CodecTypes['pg/text@1']['input'];
    };
  };
};
export type TypeMaps = TypeMapsType<
  CodecTypes,
  QueryOperationTypes,
  FieldOutputTypes,
  FieldInputTypes,
  StorageColumnTypes,
  StorageColumnInputTypes,
  AggregateTypes
>;
// ...
export type Contract = ContractWithTypeMaps<ContractBase, TypeMaps>;